Implement JarRunner
