+++ API/src/main/java/org/sikuli/script/runners/JarRunner.java
@@ -3,25 +3,24 @@
  */
 package org.sikuli.script.runners;
 
+import java.io.IOException;
+import java.util.jar.JarFile;
+import java.util.zip.ZipFile;
+
+import org.apache.commons.io.FilenameUtils;
+
 /**
  * Runs a sikulix script packed in a jar file
- *
- * NOT SUPPORTED YET!
  *
  * @author mbalmer
  *
  */
 
-public class JarRunner extends AbstractScriptRunner {
+public class JarRunner extends ZipRunner {
 
-  public static final String NAME = "RunnableJar";
+  public static final String NAME = "SikulixJar";
   public static final String TYPE = "text/jar";
   public static final String[] EXTENSIONS = new String[] {"jar"};
-
-  @Override
-  public boolean isSupported() {
-    return false;
-  }
 
   @Override
   public String getName() {
@@ -36,5 +35,15 @@ public class JarRunner extends AbstractScriptRunner {
   @Override
   public String getType() {
     return TYPE;
+  }
+
+  @Override
+  protected ZipFile openZipFile(String identifier) throws IOException {
+    return new JarFile(identifier);
+  }
+
+  @Override
+  protected String getScriptEntryName(ZipFile file) {
+    return FilenameUtils.getBaseName(file.getName()).replace("_sikuli", "") + "$py";
   }
 }
