+++ API/src/main/java/org/sikuli/script/runners/SikulixRunner.java
@@ -85,5 +85,4 @@ public class SikulixRunner extends AbstractScriptRunner {
       currentRunner.abort();
     }
   }
-
 }
